+++ plugin/src/test/java/com/stratio/cassandra/lucene/index/FSIndexTest.java
@@ -20,6 +20,7 @@ package com.stratio.cassandra.lucene.index;
 
 import com.google.common.collect.Sets;
 import com.stratio.cassandra.lucene.IndexOptions;
+import org.apache.cassandra.utils.Pair;
 import org.apache.lucene.analysis.standard.StandardAnalyzer;
 import org.apache.lucene.document.Document;
 import org.apache.lucene.document.Field;
@@ -34,6 +35,8 @@ import org.junit.rules.TemporaryFolder;
 
 import java.io.IOException;
 import java.nio.file.Paths;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Set;
 import java.util.UUID;
 
@@ -63,9 +66,10 @@ public class FSIndexTest {
                                     IndexOptions.DEFAULT_MAX_MERGE_MB,
                                     IndexOptions.DEFAULT_MAX_CACHED_MB,
                                     null);
-        index.init();
-
         Sort sort = new Sort(new SortField("field", SortField.Type.STRING));
+        Set<String> fields = Sets.newHashSet("field");
+        index.init(sort, fields);
+
         assertEquals("Index must be empty", 0, index.getNumDocs());
 
         Term term1 = new Term("field", "value1");
@@ -85,15 +89,12 @@ public class FSIndexTest {
         assertEquals("Expected 2 documents", 2, index.getNumDocs());
 
         Query query = new WildcardQuery(new Term("field", "value*"));
-        Set<String> fields = Sets.newHashSet("field");
-
-        // Search
-        //        Iterator<Document> iterator;
-        //            iterator = index.search( query, sort, null, 1, fields);
-        //            assertEquals("Expected 1 document", 1, results.size());
-        //            ScoreDoc last3 = results.values().iterator().next();
-        //            results = index.search(searcher, query, sort, last3, 1, fields);
-        //            assertEquals("Expected 1 document", 1, results.size());
+        DocumentIterator documentIterator = index.search( query, sort, null, 1);
+        List<Document> documents = new ArrayList<>();
+        while (documentIterator.hasNext()) {
+            documents.add(documentIterator.next().left);
+        }
+        assertEquals("Expected 2 documents", 2, documents.size());
 
         // Delete by term
         index.delete(term1);